literal_vs_metaphorical/requirements.txt
torch
torchvision
transformers
pillow
scikit-learn
tqdm